+++ configure.ac
@@ -35,7 +35,7 @@
 #      Makefile.in ---'                    `-> Makefile ---'
 
 AC_PREREQ([2.69])
-AC_INIT([systemf], [0.9], [systemf@patnan.com])
+AC_INIT([systemf], [0.9.0], [systemf@patnan.com])
 AC_CONFIG_AUX_DIR([build-aux])
 AM_INIT_AUTOMAKE([foreign subdir-objects]) # Does not require NEWS, COPYING, AUTHORS, ChangeLog or README
