+++ docker/Dockerfile
@@ -4,5 +4,4 @@ FROM ubuntu
 RUN apt-get update
 RUN apt-get update
 RUN apt-get install -y git build-essential autotools-dev automake libtool bison flex lcov python3 gdbserver
-COPY systemf-build systemf-gdbserv systemf-check /usr/bin/
-
+COPY systemf-build systemf-gdbserv systemf-check systemf-release-build systemf-sleep-infinity /usr/bin/
