+++ docker/docker-compose.yml
@@ -19,4 +19,4 @@ services:
       - seccomp=unconfined
     ports:
       - "${GDBSERVER_LOCAL_PORT:-2345}:${GDBSERVER_PORT:-2345}"
-    command: ["${COMMAND:-/usr/bin/systemf-gdbserv}"]
+    command: ["${COMMAND:-/usr/bin/systemf-sleep-infinity}"]
