+++ docker/systemf-sleep-infinity
@@ -1,0 +1,2 @@
+#!/bin/sh
+exec sleep infinity
